+++ src/Pim/Bundle/CatalogBundle/Entity/ProductAssociation.php
@@ -4,9 +4,10 @@ namespace Pim\Bundle\CatalogBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use JMS\Serializer\Annotation\ExclusionPolicy;
-use Pim\Bundle\CatalogBundle\Entity\Association;
+use Pim\Bundle\CatalogBundle\Entity\AssociationType;
 use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
+use Pim\Bundle\CatalogBundle\Model\ReferableInterface;
 
 /**
  * Product association entity
@@ -17,7 +18,7 @@ use Pim\Bundle\CatalogBundle\Model\ProductInterface;
  *
  * @ExclusionPolicy("all")
  */
-class ProductAssociation
+class ProductAssociation implements ReferableInterface
 {
     /**
      * @var integer
@@ -25,9 +26,9 @@ class ProductAssociation
     protected $id;
 
     /**
-     * @var Association $association
+     * @var AssociationType $associationType
      */
-    protected $association;
+    protected $associationType;
 
     /**
      * @var ProductInterface $owner
@@ -64,27 +65,27 @@ class ProductAssociation
     }
 
     /**
-     * Set association
+     * Set association type
      *
-     * @param Association $association
+     * @param AssociationType $associationType
      *
      * @return ProductAssociation
      */
-    public function setAssociation(Association $association)
+    public function setAssociationType(AssociationType $associationType)
     {
-        $this->association = $association;
+        $this->associationType = $associationType;
 
         return $this;
     }
 
     /**
-     * Get association
+     * Get association type
      *
-     * @return Association
+     * @return AssociationType
      */
-    public function getAssociation()
+    public function getAssociationType()
     {
-        return $this->association;
+        return $this->associationType;
     }
 
     /**
@@ -217,5 +218,13 @@ class ProductAssociation
         $this->groups->removeElement($group);
 
         return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getReference()
+    {
+        return $this->owner->getIdentifier() . '.' . $this->associationType->getCode();
     }
 }