+++ src/Pim/Bundle/ImportExportBundle/Reader/File/CsvReader.php
@@ -137,6 +137,7 @@ class CsvReader extends FileReader implements
     public function setUploadedFile(File $uploadedFile)
     {
         $this->filePath = $uploadedFile->getRealPath();
+        $this->csv = null;
 
         return $this;
     }
@@ -150,6 +151,7 @@ class CsvReader extends FileReader implements
     public function setFilePath($filePath)
     {
         $this->filePath = $filePath;
+        $this->csv = null;
 
         return $this;
     }
@@ -281,7 +283,9 @@ class CsvReader extends FileReader implements
             if ($data === array(null) || $data === null) {
                 return null;
             }
-            $this->stepExecution->incrementSummaryInfo('read');
+            if ($this->stepExecution) {
+                $this->stepExecution->incrementSummaryInfo('read');
+            }
 
             if (count($this->fieldNames) !== count($data)) {
                 throw new InvalidItemException(