+++ src/Pim/Bundle/ImportExportBundle/Tests/Unit/Processor/TransformerProcessorTest.php
@@ -13,7 +13,7 @@ use Pim\Bundle\ImportExportBundle\Exception\ParametrizedException;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class TransformerProcessorTest extends AbstractTransformerProcessorTestCase
+class TransformerProcessorTest extends TransformerProcessorTestCase
 {
     protected $processor;
     protected $transformer;