+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Allow to back to the grid or create another product when editing one
 - Add products to many groups through mass edit wizard
 - Attribute options fixture
+- Product associations fixture
 - Fixtures can be in CSV (all fixtures except users and currencies)
 - Fixture files can be imported through a command (all fixtures except users and currencies)
 - Add quick create popin for jobs
@@ -29,7 +30,9 @@
 - Allow to switch configuration between ORM and ODM
 - Update OroPlatform from beta-1 to beta-5
 - Move Batch Form Types to ImportExport bundle and refactor them to be able to configure any kind of job
-- Hide entity creation/deletion buttons if users don't have the corresponding rights
+- Don't display several UI elements when users don't have the corresponding rights
+- Use aliases for subforms, no more manual instanciation to enhance extensibility
+- Product prices can now be imported with a single column per currency
 
 ## Bug fixes
 - Missing pending versionable entities
@@ -74,6 +77,8 @@
 - Change attribute type configuration, refactor the attribute type compiler pass and attribute type factory
 - Remove getAttributeOptionValueRepository, getFlexibleValueRepository from FlexibleManager
 - Attribute fixtures format has changed
+- Product associations import/export format has changed.
+- Rename Association by AssociationType and all properties/methods linked to this class.
 
 # 1.0.0-beta-3 - "Hare Conditioned"